// MVCApi.AngularClient/src/api/model/productDto.ts
/**
 * MVCApi
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CategoryDto } from './categoryDto';
import { CurrencyProductDto } from './currencyProductDto';


export interface ProductDto { 
    id?: string;
    name?: string | null;
    description?: string | null;
    image?: string | null;
    categories?: Array<CategoryDto> | null;
    price?: CurrencyProductDto;
}